Configure basic routing for app
