+++ package.json
@@ -8,9 +8,10 @@
     "node-sass": "^4.12.0",
     "react": "^16.10.2",
     "react-dom": "^16.10.2",
+    "react-leaflet": "^2.5.0",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.2.0",
     "react-stars": "^2.2.5"
-    "react-leaflet": "^2.5.0"
   },
   "scripts": {
     "start": "react-scripts start",
