+++ src/pages/MapPage.js
@@ -1,0 +1,20 @@
+import React, { useState } from 'react';
+import MainMap from '../components/Map/MainMap';
+import SearchBar from '../components/SearchBar/SearchBar';
+import FilterModal from '../components/FilterModal/FilterModal';
+
+function MapPage() {
+  const [isModalOpen, setIsModalOpen] = useState(false);
+  const toggleModal = () => {
+    setIsModalOpen(s => !s);
+  };
+  return (
+    <div className="MapPage">
+      <MainMap />
+      <SearchBar toggleModal={toggleModal} />
+      <FilterModal modalData={{ isVisible: isModalOpen, toggleModal }} />
+    </div>
+  );
+}
+
+export default MapPage;
