+++ src/pages/RestaurantPageInfo.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function RestaurantInfoPage() {
+  return (
+    <div className="MapPage">
+      <h1>Restaurant info</h1>
+    </div>
+  );
+}
